requirements.txt
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
dvc>=3.0
numpy>=1.24
